chore: cleanup code and standardise client format
