+++ bbnclient/bbnclient.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/babylonlabs-io/babylon/client/query"
 	bbntypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
-	"github.com/babylonlabs-io/finality-gadget/types"
 	sdkquerytypes "github.com/cosmos/cosmos-sdk/types/query"
 )
 
@@ -13,7 +12,21 @@ type BabylonClient struct {
 	*query.QueryClient
 }
 
-var _ types.IBabylonClient = &BabylonClient{}
+var _ IBabylonClient = &BabylonClient{}
+
+//////////////////////////////
+// CONSTRUCTOR
+//////////////////////////////
+
+func NewBabylonClient(queryClient *query.QueryClient) *BabylonClient {
+	return &BabylonClient{
+		QueryClient: queryClient,
+	}
+}
+
+//////////////////////////////
+// METHODS
+//////////////////////////////
 
 func (bbnClient *BabylonClient) QueryAllFpBtcPubKeys(consumerId string) ([]string, error) {
 	pagination := &sdkquerytypes.PageRequest{}
@@ -145,6 +158,61 @@ func (bbnClient *BabylonClient) QueryFpEarliestActiveDelBtcHeight(fpPubkeyHex st
 		pagination.Key = resp.Pagination.NextKey
 	}
 	return earliestBtcHeight, nil
+}
+
+//////////////////////////////
+// INTERNAL
+//////////////////////////////
+
+// we implemented exact logic as in GetStatus
+// https://github.com/babylonlabs-io/babylon-private/blob/3d8f190c9b0c0795f6546806e3b8582de716cd60/x/btcstaking/types/btc_delegation.go#L90-L111
+func (bbnClient *BabylonClient) isDelegationActive(
+	btcDel *bbntypes.BTCDelegationResponse,
+	btcHeight uint64,
+) (bool, error) {
+	btccheckpointParams, err := bbnClient.QueryClient.BTCCheckpointParams()
+	if err != nil {
+		return false, err
+	}
+	btcstakingParams, err := bbnClient.QueryClient.BTCStakingParams()
+	if err != nil {
+		return false, err
+	}
+	kValue := btccheckpointParams.GetParams().BtcConfirmationDepth
+	wValue := btccheckpointParams.GetParams().CheckpointFinalizationTimeout
+	covQuorum := btcstakingParams.GetParams().CovenantQuorum
+	ud := btcDel.UndelegationResponse
+
+	if len(ud.GetDelegatorUnbondingSigHex()) > 0 {
+		return false, nil
+	}
+
+	// k is not involved in the `GetStatus` logic as Babylon will accept a BTC delegation request
+	// only when staking tx is k-deep on BTC.
+	//
+	// But the msg handler performs both checks 1) ensure staking tx is k-deep, and 2) ensure the
+	// staking tx's timelock has at least w BTC blocks left.
+	// (https://github.com/babylonlabs-io/babylon-private/blob/3d8f190c9b0c0795f6546806e3b8582de716cd60/x/btcstaking/keeper/msg_server.go#L283-L292)
+	//
+	// So after the msg handler accepts BTC delegation the 1st check is no longer needed
+	// the k-value check is added per
+	//
+	// So in our case, we need to check both to ensure the delegation is active
+	if btcHeight < btcDel.StartHeight+kValue || btcHeight+wValue > btcDel.EndHeight {
+		return false, nil
+	}
+
+	if uint32(len(btcDel.CovenantSigs)) < covQuorum {
+		return false, nil
+	}
+	if len(ud.CovenantUnbondingSigList) < int(covQuorum) {
+		return false, nil
+	}
+	if len(ud.CovenantSlashingSigs) < int(covQuorum) {
+		return false, nil
+	}
+
+	return true, nil
 }
 
 // The active delegation needs to satisfy:
