+++ btcclient/config.go
