+++ db/interface.go
@@ -1,0 +1,14 @@
+package db
+
+import "github.com/babylonlabs-io/finality-gadget/types"
+
+type IDatabaseHandler interface {
+	CreateInitialSchema() error
+	InsertBlock(block *types.Block) error
+	GetBlockByHeight(height uint64) (*types.Block, error)
+	GetBlockStatusByHeight(height uint64) (bool, error)
+	GetBlockStatusByHash(hash string) (bool, error)
+	GetLatestBlock() (*types.Block, error)
+	DeleteDB() error
+	Close() error
+}
