+++ finalitygadget/interface.go
@@ -1,4 +1,6 @@
-package types
+package finalitygadget
+
+import "github.com/babylonlabs-io/finality-gadget/types"
 
 type IFinalityGadget interface {
 	/* QueryIsBlockBabylonFinalized checks if the given L2 block is finalized by the Babylon finality gadget
@@ -17,7 +19,7 @@ type IFinalityGadget interface {
 	 *   - calculate voted voting power
 	 *   - check if the voted voting power is more than 2/3 of the total voting power
 	 */
-	QueryIsBlockBabylonFinalized(block *Block) (bool, error)
+	QueryIsBlockBabylonFinalized(block *types.Block) (bool, error)
 
 	/* QueryBlockRangeBabylonFinalized searches for a row of consecutive finalized blocks in the block range, and returns
 	 * the last finalized block height
@@ -33,7 +35,7 @@ type IFinalityGadget interface {
 	 * Note: caller needs to make sure the given queryBlocks are consecutive (we don't check hashes inside this method)
 	 * and start from low to high
 	 */
-	QueryBlockRangeBabylonFinalized(queryBlocks []*Block) (*uint64, error)
+	QueryBlockRangeBabylonFinalized(queryBlocks []*types.Block) (*uint64, error)
 
 	/* QueryBtcStakingActivatedTimestamp returns the timestamp when the BTC staking is activated
 	 *
@@ -54,7 +56,7 @@ type IFinalityGadget interface {
 	QueryBtcStakingActivatedTimestamp() (uint64, error)
 
 	// InsertBlock inserts a btc finalized block into the local db
-	InsertBlock(block *Block) error
+	InsertBlock(block *types.Block) error
 
 	// GetBlockStatusByHeight returns the btc finalization status of a block at given height by querying the local db
 	GetBlockStatusByHeight(height uint64) (bool, error)
@@ -63,5 +65,5 @@ type IFinalityGadget interface {
 	GetBlockStatusByHash(hash string) (bool, error)
 
 	// GetLatestBlock returns the latest finalized block by querying the local db
-	GetLatestBlock() (*Block, error)
+	GetLatestBlock() (*types.Block, error)
 }
