--- finalitygadget/utils.go
@@ -1,7 +1,0 @@
-package finalitygadget
-
-import "github.com/ethereum/go-ethereum/common"
-
-func normalizeBlockHash(hash string) string {
-	return common.HexToHash(hash).Hex()
-}
